Fix the ULibfuzzer found bug in UsbMassBoot.c file
The usb mass storage driver read the device blocksize capacity value directly
from the usb hardware. The driver assume the blocksize value cannot be 1 type.
But if a device deliberately report the blocksize as 1, which could happen in
a broken or malicious usb device, the usb mass driver will deedloop
forever when try to read or write the device.